Refactor entity IDs to Long and update DTO usage
Changed all entity ID fields from int to Long for consistency and scalability. Updated CreatePedidoDTO to use UsuarioDTO instead of a raw user ID. Added conversion methods and improved service layer DTO/entity mapping, including DetallePedidoService and PedidoService. Also added a new constructor to Usuario and started implementing convertToEntity in UsuarioService.

+++ src/main/java/org/example/topdeckapi/src/model/Categoria.java
@@ -15,7 +15,7 @@ public class Categoria {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name = "id_categoria")
-    private int idCategoria;
+    private Long idCategoria;
 
     @Column(name = "nombre")
     private String nombre;
