+++ src/main/java/org/example/topdeckapi/src/model/Pedido.java
@@ -19,7 +19,7 @@ public class Pedido {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pedido")
-    private int idPedido;
+    private Long idPedido;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_usuario", nullable = false)
